try splitting copy parser into new crate
