+++ src/pgcopy/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "pgcopy"
+description = "Library for parsing PostgreSQL COPY formatted data"
+version = "0.0.0"
+edition = "2021"
+publish = false
+rust-version = "1.58.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bytes = "1.1.0"
+csv = "1.1.6"
+pgrepr = { path = "../pgrepr" }
+postgres = { git = "https://github.com/MaterializeInc/rust-postgres", branch = "mz-0.7.2" }
+repr = { path = "../repr" }
+sql = { path = "../sql" }
