+++ src/pgwire/src/lib.rs
@@ -28,7 +28,6 @@ mod metrics;
 mod protocol;
 mod server;
 
-pub use message::CopyTextFormatParser;
 pub use metrics::Metrics;
 pub use protocol::match_handshake;
 pub use server::{Config, Server, TlsConfig, TlsMode};
